% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql-add-data.R
\name{psql_add_data}
\alias{psql_add_data}
\title{Add data to database}
\usage{
psql_add_data(
  tbl,
  schema = "public",
  tbl_name = NULL,
  config_path = getOption("psql.config_path", NULL),
  config_value = getOption("psql.value", NULL)
)
}
\arguments{
\item{tbl}{The data frame to add to the database.}

\item{schema}{A string of the schema name. Default value is \code{"public"}.}

\item{tbl_name}{A string indicating the name of the table. Default is to use
the name of the \code{tbl} argument.}

\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option.}

\item{config_value}{A string of the name of value. The default value grabs the value
from the psql.value option.}
}
\description{
Add data with a data frame to your PostgreSQL database. The data frame name
must match the table name in your database, if not use the \code{tbl_name}
argument to pass the table name.
}
\details{
The function open and closes its own database connection. You do not
need to close the database connection afterwards.

This function uses \code{DBI::dbAppendTable()} to add data to the database.
}
\examples{
\dontrun{
psql_add_data(outing, "creel")
psql_add_data(outing_new, "creel", "outing")
}
}
