% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_connect.R
\name{psql_connect}
\alias{psql_connect}
\title{Connect to PostgreSQL Database}
\usage{
psql_connect(
  config_filepath = getOption("psql.config", NULL),
  value = getOption("psql.value", NULL)
)
}
\arguments{
\item{config_filepath}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config option.}

\item{value}{A string of the name of value. The default value grabs the value
fomr the psql.value option.}
}
\value{
Returns a postgres connection
}
\description{
This function uses the information from a yaml configuration file that contains
the host, port, database name, username and password to make a connection
with a postgres database.
}
\details{
The function requires for a host, port, database name, username and
password is present in the config file given. The host, database name,
user name and password should be passed as string. The port is passed as an
integer.
}
\examples{
\dontrun{
psql_connect()
}

}
