% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql-create-db.R
\name{psql_create_db}
\alias{psql_create_db}
\title{Create PostgreSQL database}
\usage{
psql_create_db(
  dbname,
  config_path = getOption("psql.config_path", NULL),
  config_value = getOption("psql.config_value", "default")
)
}
\arguments{
\item{dbname}{A string of the name of the new database to create.}

\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option
and uses \code{NULL} if no value supplied.}

\item{config_value}{A string of the name of value. The default value grabs
the value from the psql.config_value option and uses \code{"default"} if no
value is supplied.}
}
\value{
TRUE (or errors).
}
\description{
Create a new PostgreSQL database.
}
\details{
The function open and closes its own database connection. You do not
need to close the database connection afterwards.
}
\examples{
\dontrun{
psql_create_db("new_database")
psql_create_db("new_database", config_path = "keys/config.yml")
}
}
