% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql-createdb.R
\name{psql_createdb}
\alias{psql_createdb}
\title{Create PostgreSQL database}
\usage{
psql_createdb(
  dbname,
  config_path = getOption("psql.config_path", NULL),
  config_value = getOption("psql.value", NULL)
)
}
\arguments{
\item{dbname}{A string of the name of the new database to create.}

\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option.}

\item{config_value}{A string of the name of value. The default value grabs the value
from the psql.value option.}
}
\value{
Returns TRUE
}
\description{
Create PostgreSQL database
}
\examples{
\dontrun{
psql_createdb("new_database")

psql_createdb("new_database", config_path = "keys/config.yml")
}
}
