% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql-execute-db.R
\name{psql_execute_db}
\alias{psql_execute_db}
\title{Execute SQL statement for PostgreSQL database}
\usage{
psql_execute_db(
  sql,
  config_path = getOption("psql.config_path", NULL),
  config_value = getOption("psql.value", NULL)
)
}
\arguments{
\item{sql}{A string of the SQL statement to execute.}

\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option.}

\item{config_value}{A string of the name of value. The default value grabs
the value from the psql.value option.}
}
\value{
A scalar numeric of the number of rows affected by the statement.
}
\description{
Execute PostgreSQL statements.
}
\details{
The function open and closes its own database connection. You do not
need to close the database connection afterwards. This function is a
wrapper on \verb{DBI::dbExecute().}
}
\examples{
\dontrun{
psql_execute_db(
  "CREATE SCHEMA boat_count"
)
psql_execute_db(
  "CREATE TABLE boat_count.input (
  file_name TEXT NOT NULL,
  comment TEXT)"
)
}
}
