% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql-list-tables.R
\name{psql_list_tables}
\alias{psql_list_tables}
\title{List tables in a schema}
\usage{
psql_list_tables(
  schema = "public",
  config_path = getOption("psql.config_path", NULL),
  config_value = getOption("psql.value", NULL)
)
}
\arguments{
\item{schema}{A string of the schema name. Default value is \code{"public"}.}

\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option.}

\item{config_value}{A string of the name of value. The default value grabs
the value from the psql.value option.}
}
\value{
A vector of table names
}
\description{
This function lists all the tables in a schema.
}
\details{
The function open and closes its own database connection. You do not
need to close the database connection afterwards.
}
\examples{
\dontrun{
psql_list_tables(
  "boat_count"
)
psql_list_tables()
}
}
