% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql-read-table.R
\name{psql_read_table}
\alias{psql_read_table}
\title{Read a table from a PostgreSQL database}
\usage{
psql_read_table(
  tbl_name,
  schema = "public",
  config_path = getOption("psql.config_path", NULL),
  config_value = getOption("psql.value", NULL)
)
}
\arguments{
\item{tbl_name}{A string indicating the name of the table.}

\item{schema}{A string of the schema name. Default value is \code{"public"}.}

\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option.}

\item{config_value}{A string of the name of value. The default value grabs
the value from the psql.value option.}
}
\value{
A data frame
}
\description{
Read a table from a PostgreSQL database as a data frame.
}
\details{
The function open and closes its own database connection. You do not
need to close the database connection afterwards.
}
\examples{
\dontrun{
psql_read_table("capture")
psql_read_table("counts", "boat_count")
}
}
