% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters for readwriteaws functions}
\arguments{
\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option.}

\item{config_value}{A string of the name of value. The default value grabs the value
from the psql.value option.}

\item{tbl_name}{A string indicating the name of the table.}

\item{schema}{A string of the schema name. Default value is \code{"public"}.}
}
\description{
Descriptions of the parameters for readwriteaws functions
}
\keyword{internal}
