% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql-add-data.R
\name{psql_add_data}
\alias{psql_add_data}
\title{Title}
\usage{
psql_add_data(
  tbl,
  schema = "public",
  tbl_name = NULL,
  config_path = getOption("psql.config_path", NULL),
  config_value = getOption("psql.value", NULL)
)
}
\arguments{
\item{tbl}{A dataframe that}

\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option.}

\item{config_value}{A string of the name of value. The default value grabs the value
from the psql.value option.}
}
\description{
Title
}
