% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql-backup.R
\name{psql_backup}
\alias{psql_backup}
\title{Save PostgreSQL backup}
\usage{
psql_backup(
  path = "dump_db.sql",
  config_path = getOption("psql.config_path", NULL),
  config_value = getOption("psql.value", NULL)
)
}
\arguments{
\item{path}{A string of the file path to save the dumped database}

\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option.}

\item{config_value}{A string of the name of value. The default value grabs the value
from the psql.value option.}
}
\value{
Returns TRUE
}
\description{
Save a copy of your database in a plain text format. This saves all the
SQL code to recreate the structure and data.
}
\details{
This function requires the user has postgres installed on their
machine.

Password protected databases require an extra step. A
\href{https://www.postgresql.org/docs/current/libpq-pgpass.html}{.pgpass} file
needs to be created so you the password can be passed to the function.
}
\examples{
\dontrun{
psql_backup("/Users/user1/Dumps/dump_db.sql")
}
}
