% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_connect.R
\name{psql_connect}
\alias{psql_connect}
\title{Connect to PostgreSQL Database}
\usage{
psql_connect(
  config_path = getOption("psql.config", NULL),
  value = getOption("psql.value", NULL)
)
}
\arguments{
\item{config_path}{A string of a file path to the yaml configuration file.
The default value grabs the file path from the psql.config_path option.}

\item{value}{A string of the name of value. The default value grabs the value
from the psql.value option.}
}
\value{
Returns a postgres connection
}
\description{
Connect to the default local database or a specific database by supplying
the file path for a config.yml file.
}
\details{
The yml file will use contain connection details for: host, port,
dbname, user and password. The host, database name, user name and password
should be passed as string. The port is passed as an integer.
The function uses \code{config::get()} to read the config file, check there for
more information on reading in the config data.

If no config details are passed through it will connect to your local
cluster.
}
\examples{
\dontrun{
psql_connect()
psql_connect(config_path = "config.yml", value = "database")
}

}
